# docs/API/openapi.yaml
openapi: 3.0.0
info:
  title: AI Assistant API
  version: 1.0.0
  description: |
    RAG-based policy assistant API for Caribbean government staff.

    ## Authentication
    All endpoints except `/branding` require authentication via NextAuth session cookies.

    ## User Roles
    - **admin**: Full system access
    - **superuser**: Manage documents/users in assigned categories
    - **user**: Query documents in subscribed categories
  contact:
    name: AI Assistant Support
  license:
    name: Proprietary
servers:
  - url: https://ai.abhirup.app/api
    description: Production
  - url: http://localhost:3000/api
    description: Local Development
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          enum:
            - AUTH_REQUIRED
            - ADMIN_REQUIRED
            - SUPERUSER_REQUIRED
            - NOT_FOUND
            - FORBIDDEN
            - INVALID_BODY
            - INVALID_ID
            - INVALID_DAYS
            - VALIDATION
            - VALIDATION_ERROR
            - FILE_TOO_LARGE
            - UPLOAD_LIMIT
            - INVALID_FILE_TYPE
            - DUPLICATE
            - INTERNAL
            - SERVICE_ERROR
            - LEGACY_WRITE_DISABLED
        details:
          type: string
          description: Additional context
      required:
        - error
...
    AiSetupOrganization:
      type: object
      description: Organization visible through the consolidated AI & API Setup surface.
      required:
        - id
        - name
        - type
        - isDefault
        - credentialMode
        - status
        - isolationMode
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          example: 42
        name:
          type: string
          maxLength: 120
          example: Ministry of Finance
        type:
          type: string
          enum: [DEFAULT, ENTITY, INDIVIDUAL]
        isDefault:
          type: boolean
        credentialMode:
          type: string
          enum: [PLATFORM_MANAGED, ORGANIZATION_BYOK]
        status:
          type: string
          example: active
        isolationMode:
          type: string
          example: SOFT
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        activeCredentialCount:
          type: integer
          minimum: 0
          description: Included on organizations in the consolidated overview.
        membershipRole:
          type: string
          nullable: true
          enum: [org_admin, member]
          description: The viewer's role in this organization, if applicable.
...
    AiSetupMember:
      type: object
      required: [userId, email, name, role, status]
      properties:
        userId:
          type: integer
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [org_admin, member]
        status:
          type: string
          example: active
...
    AiSetupUsageByCredential:
      type: object
      required: [credentialId, providerId, totalTokens, callCount, totalCost]
      properties:
        credentialId:
          type: string
          nullable: true
        providerId:
          type: string
          nullable: true
        totalTokens:
          type: integer
          minimum: 0
        callCount:
          type: integer
          minimum: 0
        totalCost:
          type: number
          format: double
          minimum: 0
          description: Zero when cost visibility is denied.
    AiSetupUsageSummary:
      type: object
      required: [totalTokens, totalCalls, totalCost, byCredential, costUnavailable]
      properties:
        totalTokens:
          type: integer
          minimum: 0
        totalCalls:
          type: integer
          minimum: 0
        totalCost:
          type: number
          format: double
          minimum: 0
          description: Zero when cost visibility is denied.
        byCredential:
          type: array
          items:
            $ref: '#/components/schemas/AiSetupUsageByCredential'
        costUnavailable:
          type: boolean
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key for agent bot invocation
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Unauthorized"
            code: "AUTH_REQUIRED"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Admin access required"
            code: "ADMIN_REQUIRED"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Resource not found"
            code: "NOT_FOUND"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            error: "Internal server error"
            code: "SERVICE_ERROR"
security:
  - sessionAuth: []
paths:
...
  /admin/ai-setup/organizations/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Numeric organization ID. Server-side organization scope is always enforced.
        schema:
          type: integer
    patch:
      tags:
        - Admin - AI & API Setup
      summary: Update an organization
      description: |
        Updates the name and/or credential mode. Requires `super_admin`, or a global
        `admin` whose server-resolved membership is `org_admin` for this organization.
      operationId: updateAiSetupOrganization
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              minProperties: 1
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                credentialMode:
                  type: string
                  enum: [PLATFORM_MANAGED, ORGANIZATION_BYOK]
      responses:
        '200':
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiSetupOrganization'
        '400':
          description: Invalid ID, body, name, credential mode, or empty update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/ai-setup/organizations/{id}/members:
    parameters:
      - name: id
        in: path
        required: true
        description: Numeric organization ID. Server-side organization scope is always enforced.
        schema:
          type: integer
    get:
      tags:
        - Admin - AI & API Setup
      summary: List organization members
      description: "Requires `super_admin` or a global `admin` with access to its own organization."
      operationId: listAiSetupMembers
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Organization member list
          content:
            application/json:
              schema:
                type: object
                required: [members]
                properties:
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/AiSetupMember'
        '400':
          description: Invalid organization ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Admin - AI & API Setup
      summary: Add a member or delegate organization administration
      description: |
        Adds/reactivates a user by numeric `userId` or existing-user `email`. Setting
        `role` to `org_admin` delegates administration and demotes the prior active
        administrator so exactly one remains. `INDIVIDUAL` organizations reject
        additional members, and the last active administrator cannot be silently
        demoted. Requires `super_admin` or the organization's server-resolved `org_admin`.
      operationId: addAiSetupMember
      security:
        - sessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [org_admin, member]
                  default: member
              description: "Supply `userId` or `email`. If both are supplied, `userId` is used."
      responses:
        '201':
          description: Member added/reactivated and role assigned
          content:
            application/json:
              schema:
                type: object
                required: [ok, userId, role]
                properties:
                  ok:
                    type: boolean
                    example: true
                  userId:
                    type: integer
                  role:
                    type: string
                    enum: [org_admin, member]
        '400':
          description: Invalid input, individual-organization member constraint, or last-admin invariant violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /admin/ai-setup/organizations/{id}/usage:
    parameters:
      - name: id
        in: path
        required: true
        description: Numeric organization ID. Server-side organization scope is always enforced.
        schema:
          type: integer
    get:
      tags:
        - Admin - AI & API Setup
      summary: Get organization-scoped token usage and cost
      description: |
        Returns usage attributed to the organization and credential for the selected
        interval. `days` defaults to 30, rejects non-numeric input, and clamps numeric
        values to 1–365. Visibility is server-enforced: `super_admin` can view all costs;
        an `org_admin` can view cost only for its own `ORGANIZATION_BYOK` organization
        with an active credential. Otherwise cost values are zeroed and `cost.reason`
        explains the denial or missing credential.
      operationId: getAiSetupOrganizationUsage
      security:
        - sessionAuth: []
      parameters:
        - name: days
          in: query
          required: false
          description: Number of trailing days. Numeric values are clamped to 1–365.
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Organization usage and visibility-filtered cost
          content:
            application/json:
              schema:
                type: object
                required: [org, cost, usage]
                properties:
                  org:
                    $ref: '#/components/schemas/AiSetupOrganization'
                  cost:
                    type: object
                    required: [canView, reason]
                    properties:
                      canView:
                        type: boolean
                      reason:
                        type: string
                        enum: [platform, org, denied, byok_missing_credential]
                  usage:
                    $ref: '#/components/schemas/AiSetupUsageSummary'
        '400':
          description: "Invalid organization ID or non-numeric `days`"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ========================================
  # LLM Management APIs
  # ========================================
...